pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Spaceginner", email = "ivan.demian2009@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-cors-headers (==4.7.0)",
    "django-phonenumber-field (>=8.1.0,<9.0.0)",
    "phonenumberslite (>=9.0.10,<10.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
